+++ src/screens/MainScreen/NewTaskScreen.tsx
@@ -198,16 +198,7 @@ const NewTaskScreen: React.FC<Prop> = (props: Prop) => {
             </View>
           </View>
         ) : (
-          <View
-            style={[
-              stylesCentral.center,
-              {flex: 1, backgroundColor: colors.grayBg, padding: 8},
-            ]}>
-            <Image
-              source={image.blankTask}
-              style={{width: normalize(136), height: normalize(111)}}
-            />
-            <Text style={stylesCentral.blankFont}>ยังไม่มีงานที่ต้องทำ</Text>
+          dronerStatus === 'ACTIVE' && <View>
           </View>
         )}
         {data.length == 0 && isOpenReceiveTask && dronerStatus === 'ACTIVE' && (
@@ -387,24 +378,17 @@ const NewTaskScreen: React.FC<Prop> = (props: Prop) => {
                   flexDirection: 'row',
                   justifyContent: 'space-between',
                 }}>
-                <AnimatedCircularProgress
-                  size={50}
-                  width={5}
-                  fill={Number(percentSuccess)}
-                  tintColor="#FB8705"
-                  backgroundColor="#FFEFDD"
-                  rotation={0}>
-                  {fill => (
-                    <Text
-                      style={{
-                        fontFamily: fonts.medium,
-                        fontSize: normalize(12),
-                        color: colors.fontBlack,
-                      }}>
-                      {percentSuccess + '%'}
-                    </Text>
-                  )}
-                </AnimatedCircularProgress>
+              <Image source={
+                (Number(percentSuccess) === 50)?
+                image.inprogress50:(
+                  Number(percentSuccess) === 75?
+                  image.inprogress75:
+                  image.inprogress100
+                )
+              } style={{
+                width : normalize(50),
+                height : normalize(50)
+              }}/>
                 <TouchableOpacity
                   onPress={() =>
                     navigation('MyProfileScreen', MyProfileScreen)